Fix passing of an empty argument list to XML-RPC methods.
Apparently, some servers don't like it when we pass no <params> with a
parameter-less method call. The XML-RPC spec says this:

"If the procedure call has parameters, the <methodCall> must contain a
<params> sub-item. The <params> sub-item can contain any number of
<param>s, each of which has a <value>."

I thought this means that method calls of methods without parameters
don't need <params>. Apparently, this is not the interpretation
favored by the authors of aria2c.